# RBAC: Fix slow user permission search query on MySQL