--- a/pkg/services/accesscontrol/database/database.go
+++ b/pkg/services/accesscontrol/database/database.go
@@ -11,36 +11,29 @@ import (
 )
 
 const (
-	// userPermissionsSQL is a query to select all users permissions directly assigned to them.
-	userPermissionsSQL = `SELECT ur.user_id, p.action, p.scope, ur.role_id
-	FROM permission AS p
-	INNER JOIN user_role AS ur on ur.role_id = p.role_id
-	WHERE (ur.org_id = ? OR ur.org_id = ?)`
+	// userAssignsSQL is a query to select all users permissions directly assigned to them.
+	userAssignsSQL = `SELECT ur.user_id, ur.org_id, ur.role_id
+	FROM user_role AS ur`
 
-	// teamPermissionsSQL is a query to select all users permissions associated to their team membership.
-	teamPermissionsSQL = `SELECT tm.user_id, p.action, p.scope, tr.role_id
-	FROM permission AS p
-	INNER JOIN team_role AS tr ON tr.role_id = p.role_id
-	INNER JOIN team_member AS tm ON tm.team_id = tr.team_id
-	WHERE (tr.org_id = ? OR tr.org_id = ?)`
+	// teamAssignsSQL is a query to select all users permissions associated to their team membership.
+	teamAssignsSQL = `SELECT tm.user_id, tr.org_id, tr.role_id
+	FROM team_role AS tr 
+	INNER JOIN team_member AS tm ON tm.team_id = tr.team_id`
 
-	// basicRolePermissionsSQL is a query to select all users permissions associated to their basic role (Admin, Editor, Viewer, None).
-	basicRolePermissionsSQL = `SELECT ou.user_id, p.action, p.scope, br.role_id
-	FROM permission AS p
-	INNER JOIN builtin_role AS br ON br.role_id = p.role_id
-	INNER JOIN org_user AS ou ON ou.role = br.role
-	WHERE (ou.org_id = ? OR ou.org_id = ?)`
+	// basicRoleAssignsSQL is a query to select all users permissions associated to their basic role (Admin, Editor, Viewer, None).
+	basicRoleAssignsSQL = `SELECT ou.user_id, ou.org_id, br.role_id
+	FROM builtin_role AS br
+	INNER JOIN org_user AS ou ON ou.role = br.role`
 
-	// grafanaAdminPermissionsSQL is a query to select all grafana admin users permissions
+	// grafanaAdminAssignsSQL is a query to select all grafana admin users permissions
 	// it has to be formatted with the quoted user table.
-	grafanaAdminPermissionsSQL = `SELECT sa.user_id, p.action, p.scope, br.role_id
-	FROM permission AS p
-	INNER JOIN builtin_role AS br ON br.role_id = p.role_id
+	grafanaAdminAssignsSQL = `SELECT sa.user_id, br.org_id, br.role_id
+	FROM builtin_role AS br
 	INNER JOIN (
 		SELECT u.id AS user_id
 	    FROM %s AS u WHERE u.is_admin
 	) AS sa ON 1 = 1
-	WHERE br.role = ? AND (br.org_id = ? OR br.org_id = ?)`
+	WHERE br.role = ?`
 )
 
 func ProvideService(sql db.DB) *AccessControlStore {
@@ -87,35 +80,6 @@ func (s *AccessControlStore) GetUserPermissions(ctx context.Context, query acces
 	return result, err
 }
 
-func searchSubqueryWithFilters(query string, table string, orgID int64, params []any, options accesscontrol.SearchOptions) (string, []any, error) {
-	params = append(params, orgID, accesscontrol.GlobalOrgID)
-	if options.NamespacedID != "" {
-		query += " AND " + table + ".user_id = ?"
-		userID, err := options.ComputeUserID()
-		if err != nil {
-			return "", nil, err
-		}
-		params = append(params, userID)
-	}
-	if options.ActionPrefix != "" {
-		query += ` AND p.action LIKE ?`
-		params = append(params, options.ActionPrefix+"%")
-	}
-	if options.Action != "" {
-		query += ` AND p.action = ?`
-		params = append(params, options.Action)
-	}
-	if options.Scope != "" {
-		// Search for scope and wildcard that include the scope
-		scopes := append(options.Wildcards(), options.Scope)
-		query += ` AND p.scope IN ( ? ` + strings.Repeat(", ?", len(scopes)-1) + ")"
-		for i := range scopes {
-			params = append(params, scopes[i])
-		}
-	}
-	return query, params, nil
-}
-
 // SearchUsersPermissions returns the list of user permissions in specific organization indexed by UserID
 func (s *AccessControlStore) SearchUsersPermissions(ctx context.Context, orgID int64, options accesscontrol.SearchOptions) (map[int64][]accesscontrol.Permission, error) {
 	type UserRBACPermission struct {
@@ -125,42 +89,54 @@ func (s *AccessControlStore) SearchUsersPermissions(ctx context.Context, orgID i
 	}
 	dbPerms := make([]UserRBACPermission, 0)
 
+	var userID int64
+	if options.NamespacedID != "" {
+		var err error
+		userID, err = options.ComputeUserID()
+		if err != nil {
+			return nil, err
+		}
+	}
+
 	if err := s.sql.WithDbSession(ctx, func(sess *db.Session) error {
 		roleNameFilterJoin := ""
 		if len(options.RolePrefixes) > 0 {
-			roleNameFilterJoin = "INNER JOIN role AS r on up.role_id = r.id"
+			roleNameFilterJoin = "INNER JOIN role AS r ON up.role_id = r.id"
 		}
 
 		params := []any{}
 
-		direct, params, err := searchSubqueryWithFilters(userPermissionsSQL, "ur", orgID, params, options)
-		if err != nil {
-			return err
+		direct := userAssignsSQL
+		if options.NamespacedID != "" {
+			direct += " WHERE ur.user_id = ?"
+			params = append(params, userID)
 		}
 
-		team, params, err := searchSubqueryWithFilters(teamPermissionsSQL, "tm", orgID, params, options)
-		if err != nil {
-			return err
+		team := teamAssignsSQL
+		if options.NamespacedID != "" {
+			team += " WHERE tm.user_id = ?"
+			params = append(params, userID)
 		}
 
-		basic, params, err := searchSubqueryWithFilters(basicRolePermissionsSQL, "ou", orgID, params, options)
-		if err != nil {
-			return err
+		basic := basicRoleAssignsSQL
+		if options.NamespacedID != "" {
+			basic += " WHERE ou.user_id = ?"
+			params = append(params, userID)
 		}
 
+		grafanaAdmin := fmt.Sprintf(grafanaAdminAssignsSQL, s.sql.Quote("user"))
 		params = append(params, accesscontrol.RoleGrafanaAdmin)
-		grafanaAdmin := fmt.Sprintf(grafanaAdminPermissionsSQL, s.sql.Quote("user"))
-		grafanaAdmin, params, err = searchSubqueryWithFilters(grafanaAdmin, "sa", orgID, params, options)
-		if err != nil {
-			return err
+		if options.NamespacedID != "" {
+			grafanaAdmin += " AND sa.user_id = ?"
+			params = append(params, userID)
 		}
 
 		// Find permissions
 		q := `
 		SELECT
 			user_id,
-			action,
-			scope
+			p.action,
+			p.scope
 		FROM (
 			` + direct + `
 			UNION ALL
@@ -169,12 +145,30 @@ func (s *AccessControlStore) SearchUsersPermissions(ctx context.Context, orgID i
 			` + basic + `
 			UNION ALL
 			` + grafanaAdmin + `
-		) AS up ` + roleNameFilterJoin
+		) AS up ` + roleNameFilterJoin + `
+		INNER JOIN permission AS p ON up.role_id = p.role_id
+		WHERE (up.org_id = ? OR up.org_id = ?)
+		`
+		params = append(params, orgID, accesscontrol.GlobalOrgID)
 
-		// Apply role prefix filter after the union won't change performance
-		// Since in OSS all roles should match and in ENT we do not use it
+		if options.ActionPrefix != "" {
+			q += ` AND p.action LIKE ?`
+			params = append(params, options.ActionPrefix+"%")
+		}
+		if options.Action != "" {
+			q += ` AND p.action = ?`
+			params = append(params, options.Action)
+		}
+		if options.Scope != "" {
+			// Search for scope and wildcard that include the scope
+			scopes := append(options.Wildcards(), options.Scope)
+			q += ` AND p.scope IN ( ? ` + strings.Repeat(", ?", len(scopes)-1) + ")"
+			for i := range scopes {
+				params = append(params, scopes[i])
+			}
+		}
 		if len(options.RolePrefixes) > 0 {
-			q += " WHERE ( " + strings.Repeat("r.name LIKE ? OR ", len(options.RolePrefixes)-1)
+			q += " AND ( " + strings.Repeat("r.name LIKE ? OR ", len(options.RolePrefixes)-1)
 			q += "r.name LIKE ? )"
 			for _, prefix := range options.RolePrefixes {
 				params = append(params, prefix+"%")