**What is this feature?**

We have performance issues with the search users permissions query when filtering on userID. It seems like mysql  does not apply the overall `where clause` to subqueries. This PR intends to fix this.

In this PR, we first make the union of all users/teams/basic role/server admin assignments by user, then join with the permission table to apply the action & scope filtering. 

I ran the benchmarks locally before the change and after the change:
* `BenchmarkSearchUsersWithActionPrefix_1K_1K`:		
  ||before (ns/op)|after (ns/op)|percent
  |-- | -- | -- | --|
  |mysql | `11103623352` | `3475177081` | `31.29768519`|
  |postgres | `3036236726` | `2901494590` | `95.56219926`|
  |sqlite3 | `4166466507` | `4054950690` | `97.32349182`|
  
  Interestingly, the new approach is more efficient when searching all users permissions by Action Prefix on MySQL.
* `BenchmarkSearchUsersWithPerm_1K_1K`:
  || before (ns/op) | after (ns/op) | percent |
  |-- | -- | -- | --|  
  |mysql | `15369102335` | `16835394` | `0.109540516` |
  |postgres | `8881909` | `9293961` | `104.6392279` |
  |sqlite3 | `51378867` | `50724036` | `98.72548571`|
  
  Again, the new approach is much more efficient when searching all users permissions for users with a given action and scope.
* `BenchmarkSearchUserWithAction_1K_1k`:
  || before (ns/op)| after (ns/op)| percent|
  |--|--|--|--|
  |mysql | `11087799395` | `5417383` | `0.04885895575`|
  |postgres | `256819198` | `3949197` | `1.53773434`|
  |sqlite3 | `63115417` | `65489960` | `103.7622234`|

  Finally, this is the issue that I wanted to fix, the new approach is much more efficient on both postgres and mysql when searching for a given user permissions by action.

**Which issue(s) does this PR fix?**:

<!--

- Automatically closes linked issue when the Pull Request is merged.

Usage: "Fixes #<issue number>", or "Fixes (paste link of issue)"

-->

Fixes #

**Special notes for your reviewer:**

Please check that:
- [ ] It works as expected from a user's perspective.
- [ ] If this is a pre-GA feature, it is behind a feature toggle.
- [ ] The docs are updated, and if this is a [notable improvement](https://grafana.com/docs/writers-toolkit/writing-guide/contribute-release-notes/#how-to-determine-if-content-belongs-in-a-whats-new-document), it's added to our [What's New](https://grafana.com/docs/writers-toolkit/writing-guide/contribute-release-notes/) doc.